flask_app/requirements.txt
numpy==1.14.5
Flask==1.0.2
Keras==2.1.5
tensorflow==1.7.0
Werkzeug==0.14.1
Pillow==5.2.0
